Replace floats in header with uint32_t

+++ include/EVT/io/PWM.hpp
@@ -6,7 +6,7 @@
 namespace EVT::core::IO {
 
 // Forward declarations:
-// The diferent pins are hardware specific. Forward declarations to allow
+// The different pins are hardware specific. Forward declarations to allow
 // at compilation time the decision of which pins should be used.
 enum class Pin;
 
@@ -25,21 +25,21 @@ public:
      *
      * @param dutyCycle[in] Duty cycle to set the pin to.
      */
-    virtual void setDutyCycle(float dutyCycle) = 0;
+    virtual void setDutyCycle(uint32_t dutyCycle) = 0;
 
     /**
      * Set the period for the PWM in seconds.
      *
      * @param period[in] The period of the PWM in seconds.
      */
-    virtual void setPeriod(float period) = 0;
+    virtual void setPeriod(uint32_t period) = 0;
 
     /**
      * Get the current duty cycle.
      *
      * @return The duty cycle the PWM is operating at.
      */
-    virtual float getDutyCycle() = 0;
+    virtual uint32_t getDutyCycle() = 0;
 
     /**
      * Get the current period.
@@ -52,7 +52,7 @@ protected:
     /// The pin the PWM is attached to
     Pin pin;
     /// The duty cycle of the PWM
-    float dutyCycle;
+    uint32_t dutyCycle;
     /// The period of the PWM
     uint32_t period;
 };
